Created new db (same parameters) modified dockerfile for keeping alive docker test refined, streamlit dashboard renamed to dashboard
Created new db (same parameters)
modified dockerfile for keeping alive docker test refined,
 streamlit dashboard renamed to dashboard
added and tested runner_pipeline.sh (in container)
