+++ docs/environment_for_docker.txt
@@ -80,3 +80,7 @@ apt-get update && apt-get install -y liblzma5 <- this can added to dockerfile as
     curl \
     pkg-config \
     && rm -rf /var/lib/apt/lists/*
+
+   PyMuPDF instead of fitz
+
+   docker cp test_full.py linux_with_cuda-complete-container:/app/01_RAG/tests
