+++ src/runner_dashboard.bat
@@ -8,4 +8,4 @@ echo ====================================
 
 echo.
 echo Press Ctrl+C to stop the application when done.
-%STREAMLIT% run streamlit_modern_multiuser.py
+%STREAMLIT% run streamlit.py
