+++ tests/test_inference.py
@@ -8,6 +8,7 @@ Usage:
     python test_inference.py --model InternVL3_5-2B-Q6_K --mode all
     python test_inference.py --model Qwen3-8B-Q4_K_M --mode quick
     python test_inference.py --model Qwen3-8B-Q5_K_M --mode all
+    python3 test_inference.py --model Qwen3-8B-Q5_K_M --mode all
     python test_inference.py --model InternVL3_5-2B-Q6_K --mode single --question-id 1
     python test_inference.py --show-stats
     python test_inference.py --export-report
